@@ -29,46 +29,13 @@
 #
 #======================================================================
 
-Rails.application.configure do
-  config.cache_classes = true
-  config.eager_load = false
-  config.consider_all_requests_local = true
-  config.serve_static_files = true
-  config.force_ssl = false
-  config.autoflush_log = false
+Rails.application.config.tap do |config|
+  config.session_store :cookie_store, {
+    key: 'hawk',
 
-  config.action_dispatch.show_exceptions = false
-  config.action_dispatch.cookies_serializer = :json
-
-  config.action_controller.perform_caching = false
-  config.action_controller.allow_forgery_protection = false
-
-  # config.action_mailer.raise_delivery_errors = false
-  # config.action_mailer.delivery_method = :test
-
-  config.action_view.raise_on_missing_translations = true
-
-  config.active_support.deprecation = :stderr
-
-  # config.active_record.migration_error = :page_load
-  # config.active_record.dump_schema_after_migration = false
-
-  config.assets.debug = false
-  config.assets.raise_runtime_errors = true
-  config.assets.js_compressor = nil
-  config.assets.css_compressor = nil
-  config.assets.compile = true
-  config.assets.digest = false
-  config.assets.manifest = Rails.root.join("public", "assets", "manifest.json")
-
-  config.i18n.fallbacks = true
-
-  config.log_level = :debug
-  config.log_tags = []
-
-  # config.logger = ActiveSupport::TaggedLogging.new(
-  #   Logger.new(Rails.root.join("log", "test.log"))
-  # )
-
-  # config.logger.formatter = ::Hawk::Logger::Formatter.new
+    # Allow session cookie to persist for a (somewhat arbitrary) ten days.
+    # This means when using the dashboard you won't be required to log in
+    # to all your clusters all the time.
+    expire_after: 60 * 60 * 24 * 10
+  }
 end
